move rotate funcs to sim/utils
