+++ neurodsp/tests/spectral/test_utils.py
@@ -40,18 +40,3 @@ def test_trim_spectrogram():
     f_ext, t_ext, p_ext = trim_spectrogram(freqs, times, pows, f_range=[6, 8], t_range=None)
     assert_equal(f_ext, np.array([6, 7, 8]))
     assert_equal(t_ext, times)
-
-def test_rotate_timeseries(tsig):
-
-    out = rotate_timeseries(tsig, FS, 0.25)
-    assert out.shape == tsig.shape
-    assert not np.array_equal(out, tsig)
-
-def test_rotate_powerlaw():
-
-    freqs = np.array([5, 6, 7, 8, 9])
-    pows = np.array([1, 2, 3, 4, 5])
-    d_exp = 1
-
-    pows_new = rotate_powerlaw(freqs, pows, d_exp)
-    assert pows.shape == pows_new.shape
